IPO Scoop Group Project #2
This is the final draft Jupyter notebook for the IPO Scoop project. This notebook contains the application and evaluation of several machine learning models to predict the profitability of investment in Initial Public Offerings based on the companies published financial data and other related factors. Potential Realized Profits/Losses were plotted according to model predictions comparing the efficacy of these models. The top model selected was also plotted with a similar 'Dollar Cost Averaging' strategy for investment in SPY ETF and returns were compared.